Refactor all movement sys_calls into their own file
